+++ examples/register_gamepads.rs
@@ -48,8 +48,10 @@ fn join(
                     (Action::Jump, GamepadButtonType::South),
                     (Action::Disconnect, GamepadButtonType::Select),
                 ])
-                // Make sure to set the gamepad or all gamepads will be used!
-                .set_gamepad(gamepad)
+                // Make sure to set the gamepad!
+                // If you don't want a specified exclusive gamepad,
+                // use `set_gamepad()` instead.
+                .set_exclusive_gamepad(gamepad)
                 .build();
                 let player = commands
                     .spawn(InputManagerBundle::with_map(input_map))